requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
bcrypt==3.2.0
passlib
python-jose[cryptography]
geoalchemy2
alembic
shapely
# Testing dependencies
pytest
httpx
